optimize oracle batch delete rollback
