+++ rm-datasource/src/main/java/io/seata/rm/datasource/undo/mysql/MySQLUndoDeleteExecutor.java
@@ -15,19 +15,16 @@
  */
 package io.seata.rm.datasource.undo.mysql;
 
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import io.seata.common.util.CollectionUtils;
 import io.seata.rm.datasource.ColumnUtils;
 import io.seata.rm.datasource.sql.struct.Field;
 import io.seata.rm.datasource.sql.struct.Row;
 import io.seata.rm.datasource.sql.struct.TableRecords;
+import io.seata.rm.datasource.undo.AbstractUndoDeleteExecutor;
 import io.seata.rm.datasource.undo.AbstractUndoExecutor;
 import io.seata.rm.datasource.undo.SQLUndoLog;
 import io.seata.sqlparser.util.JdbcConstants;
@@ -37,7 +34,11 @@ import io.seata.sqlparser.util.JdbcConstants;
  *
  * @author sharajava
  */
-public class MySQLUndoDeleteExecutor extends AbstractUndoExecutor {
+public class MySQLUndoDeleteExecutor extends AbstractUndoDeleteExecutor {
+    /**
+     * INSERT INTO a (x, y, z, pk) VALUES (?, ?, ?, ?),  (?, ?, ?, ?)
+     */
+    private static final String INSERT_SQL_TEMPLATE = "INSERT INTO %s (%s) VALUES %s";
 
     /**
      * Instantiates a new My sql undo delete executor.
@@ -48,62 +49,9 @@ public class MySQLUndoDeleteExecutor extends AbstractUndoExecutor {
         super(sqlUndoLog);
     }
 
-    /**
-     * INSERT INTO a (x, y, z, pk) VALUES (?, ?, ?, ?),  (?, ?, ?, ?)
-     */
-    private static final String INSERT_SQL_TEMPLATE = "INSERT INTO %s (%s) VALUES %s";
 
-    private static final int INSERT_BATCH_NUM = 500;
 
     @Override
-    public void executeOn(Connection conn) throws SQLException {
-        if (IS_UNDO_DATA_VALIDATION_ENABLE && !dataValidationAndGoOn(conn)) {
-            return;
-        }
-
-        try {
-            TableRecords undoRows = getUndoRows();
-
-            List<List<Row>> rowsDouble = CollectionUtils.cutData(undoRows.getRows(), INSERT_BATCH_NUM);
-            PreparedStatement undoPstCache = null;
-            for (List<Row> rows : rowsDouble) {
-                PreparedStatement undoPst = null;
-                if (null != undoPstCache && rows.size() == INSERT_BATCH_NUM) {
-                    undoPst = undoPstCache;
-                } else {
-                    undoPst = conn.prepareStatement(generateInsertSql(rows));
-                    undoPstCache = undoPst;
-                }
-
-                int undoIndex = 0;
-                for (Row undoRow : rows) {
-                    for (Field field : undoRow.nonPrimaryKeys()) {
-                        undoPst.setObject(++ undoIndex, field.getValue(), field.getType());
-                    }
-                    for (Field field : getOrderedPkList(getUndoRows(), undoRow, JdbcConstants.MYSQL)) {
-                        undoPst.setObject(++ undoIndex, field.getValue(), field.getType());
-                    }
-                }
-                undoPst.executeUpdate();
-            }
-
-        } catch (Exception ex) {
-            if (ex instanceof SQLException) {
-                throw (SQLException) ex;
-            } else {
-                throw new SQLException(ex);
-            }
-        }
-    }
-
-    /**
-     * Undo delete.
-     *
-     * Notice: PK is at last one.
-     * @see AbstractUndoExecutor#undoPrepare
-     *
-     * @return sql
-     */
     protected String generateInsertSql(List<Row> rows) {
         TableRecords beforeImage = sqlUndoLog.getBeforeImage();
 
@@ -121,10 +69,5 @@ public class MySQLUndoDeleteExecutor extends AbstractUndoExecutor {
         String insertValues = String.join(",", Collections.nCopies(rows.size(), insertValueOneRow));
 
         return String.format(INSERT_SQL_TEMPLATE, sqlUndoLog.getTableName(), insertColumns, insertValues);
-    }
-
-    @Override
-    protected TableRecords getUndoRows() {
-        return sqlUndoLog.getBeforeImage();
     }
 }
