+++ rm-datasource/src/main/java/io/seata/rm/datasource/undo/oracle/OracleUndoDeleteExecutor.java
@@ -15,17 +15,16 @@
  */
 package io.seata.rm.datasource.undo.oracle;
 
-import io.seata.common.exception.ShouldNeverHappenException;
-import io.seata.common.util.CollectionUtils;
 import io.seata.rm.datasource.ColumnUtils;
 import io.seata.rm.datasource.sql.struct.Field;
 import io.seata.rm.datasource.sql.struct.Row;
 import io.seata.rm.datasource.sql.struct.TableRecords;
-import io.seata.rm.datasource.undo.AbstractUndoExecutor;
+import io.seata.rm.datasource.undo.AbstractUndoDeleteExecutor;
 import io.seata.rm.datasource.undo.SQLUndoLog;
 import io.seata.sqlparser.util.JdbcConstants;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -34,12 +33,14 @@ import java.util.stream.Collectors;
  *
  * @author ccg
  */
-public class OracleUndoDeleteExecutor extends AbstractUndoExecutor {
+public class OracleUndoDeleteExecutor extends AbstractUndoDeleteExecutor {
 
     /**
-     * INSERT INTO a (x, y, z, pk) VALUES (?, ?, ?, ?)
+     * INSERT INTO a (x, y, z, pk1, pk2)
+     *   SELECT       ?, ?, ?, ?, ? from dual
+     *   UNION SELECT ?, ?, ?, ?, ? from dual
      */
-    private static final String INSERT_SQL_TEMPLATE = "INSERT INTO %s (%s) VALUES (%s)";
+    private static final String INSERT_SQL_TEMPLATE = "INSERT INTO %s (%s) %s";
 
     /**
      * Instantiates a new oracle undo delete executor.
@@ -50,14 +51,13 @@ public class OracleUndoDeleteExecutor extends AbstractUndoExecutor {
         super(sqlUndoLog);
     }
 
+
     @Override
-    protected String buildUndoSQL() {
-        TableRecords beforeImage = sqlUndoLog.getBeforeImage();
-        List<Row> beforeImageRows = beforeImage.getRows();
-        if (CollectionUtils.isEmpty(beforeImageRows)) {
-            throw new ShouldNeverHappenException("Invalid UNDO LOG");
-        }
-        Row row = beforeImageRows.get(0);
+    protected String generateInsertSql(List<Row> rows) {
+        TableRecords beforeImage = getUndoRows();
+
+        Row row = rows.get(0);
+
         List<Field> fields = new ArrayList<>(row.nonPrimaryKeys());
         fields.addAll(getOrderedPkList(beforeImage,row,JdbcConstants.ORACLE));
 
@@ -66,14 +66,11 @@ public class OracleUndoDeleteExecutor extends AbstractUndoExecutor {
         String insertColumns = fields.stream()
             .map(field -> ColumnUtils.addEscape(field.getName(), JdbcConstants.ORACLE))
             .collect(Collectors.joining(", "));
-        String insertValues = fields.stream().map(field -> "?")
-            .collect(Collectors.joining(", "));
+        String insertValueOneRow = "SELECT " + fields.stream().map(field -> "?")
+                .collect(Collectors.joining(", ")) + " FROM dual";
+        String insertValues = String.join(" UNION ", Collections.nCopies(rows.size(), insertValueOneRow));
 
         return String.format(INSERT_SQL_TEMPLATE, sqlUndoLog.getTableName(), insertColumns, insertValues);
     }
 
-    @Override
-    protected TableRecords getUndoRows() {
-        return sqlUndoLog.getBeforeImage();
-    }
 }
